+++ src/HomeWork1/HomeWork6/Jumping.java
@@ -1,5 +1,5 @@
 package HomeWork1.HomeWork6;
 
 public interface Jumping {
-  void jump(int height);
+  void jump(double height);
 }
